add TSTAMP to afl

+++ script/run-afl.sh
@@ -1,26 +1,33 @@
 #!/bin/bash
 RUN_DIR="$( cd "$(dirname "$0")" && pwd )"
 export FUZZER=afl
-INPUT=$CORPUS_REPO/$CORPUS/
-TMP=$RUN_DIR/../tmp
+export TSTAMP=$(date +%s)
+if [[ -z "$CORPUS_REPO" ]]; then
+  echo Must supply CORPUS_REPO env variable
+  exit 1
+fi
+if [[ -z "$CORPUS" ]]; then
+  echo Must supply CORPUS env variable
+  exit 2
+fi
+
+TMP=$RUN_DIR/../tmp-$TSTAMP
 mkdir -p $TMP
 rm -rf $TMP/*
-cp -r $INPUT/* $TMP/
+cp -r $CORPUS_REPO/$CORPUS/* $TMP/
+if [ $? -ne 0 ]; then
+  echo Invalid CORPUS
+  exit 3
+fi
 if [ $CORPUS == 'engineered' ]; then
   rm $TMP/ia64-unwind
   rm $TMP/mach-flags.a
   rm $TMP/nds32-attributes
 fi
 
-INPUT=$TMP
-
-docker run --rm -w /work -it -v "$RUN_DIR/..":/work -v "$INPUT":/corpus my/qsym sh -c "FUZZER=$FUZZER CORPUS=$CORPUS /work/script/campaign-afl.sh"
-
-FUZZER=$FUZZER CORPUS=$CORPUS $RUN_DIR/collect.js
-
-docker run --rm -w /work -it -v "$RUN_DIR/..":/work -v "$INPUT":/corpus my/qsym sh -c "FUZZER=$FUZZER CORPUS=$CORPUS USE_ASAN=1 /work/script/campaign-afl.sh"
-
-FUZZER=$FUZZER CORPUS=$CORPUS $RUN_DIR/collect-asan.sh
+docker run --rm -w /work -it -v "$RUN_DIR/..":/work -v "$INPUT":/corpus my/qsym sh -c "FUZZER=$FUZZER CORPUS=$CORPUS TSTAMP=$TSTAMP USE_ASAN=1 /work/script/campaign-afl.sh"
 
 rm $TMP/*
 rmdir $TMP
+
+TSTAMP=$TSTAMP FUZZER=$FUZZER CORPUS=$CORPUS $RUN_DIR/collect-asan.sh
